refactor: split main script
